requirements.txt
# Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
python-nmap>=0.7.1
requests>=2.31.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.2
